urdf files lab-corner, lab-corner with arm
file tams_corner.urdf:
model of the corner in the TAMS laboratory, with shelf and table.
not yet completed! details like the kable-canal are missing.
The origin (world) will maybe switch positions.

file tams_ur5_setup_corner.urdf:
merges the corner with the arm and set the arm to the correct position.