Changing eef to s_model_tool0
The old planner planed with the eef ee_link, this was the last frame of
the UR5. We changed the eef to the toolframe of the gripper to plan with
the hole arms setup.

In this couse the moveit assistent had some updates which were adopted.

The name of the planning group was changed to "arm", so that old code
can't use this group without changes.