Add duration parameters to controllers.yaml
Because we now use a FollowJointTrajectory action server for the
gripper, but don't use the created trajectories it can happen
that the execution lasts longer than moveit tolerates.
If that happens moveit will show an error message and the grippers
state will deviate from its true state, because moveit assumes
the execution was aborted.
For this reason we set the allowed_execution_duration_scaling
parameter a lot higher for the gripper than for the arm.